Added presentation layer. Added react components for home page to  calculate and display month to month mortgage details

+++ API/Model/MortgageRequest.cs
@@ -3,7 +3,7 @@
     public class MortgageRequest
     {
         public double loanAmount { get; set; }
-        public float annualInterestRate { get; set; }
+        public double annualInterestRate { get; set; }
         public int loanTerm { get; set; }
         public DateTime startDate { get; set; }
     }
